requirements.txt
fastapi
uvicorn
mysql-connector-python
transformers
matplotlib
tensorflow